Organize file structors
