+++ src/App.js
@@ -1,27 +1,15 @@
 /* eslint-disable react/prefer-stateless-function */
-import { Route, Routes, Link } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import Calculator from './component/Calculator';
 import './App.css';
-import Quote from './component/Quote';
-import Home from './component/Home';
+import Quote from './component/pages/Quote';
+import Home from './component/pages/Home';
+import Navbar from './component/pages/Navbar';
 
 export default function App() {
   return (
     <div className="App">
-      <nav>
-        <h1>Math Magicians</h1>
-        <ul className="navbar">
-          <li>
-            <Link to="/">Home</Link>
-          </li>
-          <li>
-            <Link to="calculator">Calculator</Link>
-          </li>
-          <li>
-            <Link to="quote">Quote</Link>
-          </li>
-        </ul>
-      </nav>
+      <Navbar />
       <Routes>
         <Route exact path="/" element={<Home />} />
         <Route path="calculator" element={<Calculator />} />
