+++ src/component/pages/Quote.jsx
